Partially address item NSDB-5
Review comment from

  http://www.ietf.org/mail-archive/web/nfsv4/current/msg10851.html

| 4.2.1.1
|   I note that LDAP already has a UUID syntax 1.3.6.1.1.16.1 and I
|   don't believe you should be defining yours as inheriting from
|   "name".

Rough consensus was to go with the standard LDAP "UUID" type rather
than inheriting from "name:"

  http://www.ietf.org/mail-archive/web/nfsv4/current/msg11207.html

Remove the fedfsUuid type, and change the fedFsFsnUuid and
fedFsFslUuid attributes to inherit from "UUID," defined in RFC 4530.

Note that LDAP does not provide any mechanism for verifying that a
UUID is of a particular variant.