Partially address item NSDB-5
Review comment from

  http://www.ietf.org/mail-archive/web/nfsv4/current/msg10851.html

| 4.1
|   fedfsNcePrefix is really a DN (not a string) and must conform to
|   DN syntax. That's made clear in the following definition, but
|   this description is misleading. I note that LDAP is still
|   basically X.500, and this informal definition is invalid in pure
|   X.500 terms. You should dispense with the notion of prefix and
|   just make this a regular DN, with a constraint that the DN will
|   be subordinate to the containerInfo entry.

In the current version of the NSDB draft, the suggested change has
already been made.  However, subsequent discussion

  http://www.ietf.org/mail-archive/web/nfsv4/current/msg11245.html

notes that there are some NSDB client implementation challenges with
using a partial DN.  Chu agrees that a DN prefix is not typical, and
that a full DN is preferred.

James Lentini remembers that the prefix concept was inherited from
an earlier NSDB design where the NCE DN was stored in junctions, and
thus space savings were required.  The NCE location attribute now
resides in the NSDB's namingContext entries, which have adequate
space for a full DN.

In addition:

  o  It is preferred that splitting and joining DNs be done on
     LDAP servers, not on clients

  o  Retrieving a full DN from the server makes the mechanics of
     constructing NSDB queries simpler

  o  Setting up an NSDB is simpler if the NCE DN can be retrieved
     from the LDAP server and placed in the NCI without alteration

  o  There is no longer a need to handle an empty NCE prefix.  This
     was a complication for some LDAP server admin UIs

Rename the fedfsNcePrefix attribute, making it the full DN of the
local NSDB Container Entry.  Update exposition in section 5.2.1 to
describe how the new attribute is used.