Partially address item NSDB-5
Review comment from

http://www.ietf.org/mail-archive/web/nfsv4/current/msg10851.html

| 4.2.1.2/3
|   IMO you should define a URL format instead of distinct
|   address/port attributes.

A few of us met during IETF 84 to discuss this issue.  The meeting
and further discussion are summarized here:

  http://www.ietf.org/mail-archive/web/nfsv4/current/msg11233.html

Rough consensus is that the information in these attributes is no
longer required.  Instead of using an LDAP URL to represent the
NSDB name and port, the attributes may be removed.

Remove the fedfsNsdbName and fedfsNsdbPort attributes.