Clean up section describing NSDBs
http://www.ietf.org/mail-archive/web/nfsv4/current/msg11233.html

During this discussion, James mentioned "proxied FSLs."  The text
he referred to describes something that the current NSDB architecture
appears not to support.

We could develop this concept further within the existing
architecture, or remove it from the current NSDB protocol in favor
of making it "future work."

Remove the mention for now to avoid inconsistent language.
Reorganize the NSDB section for clarity.