Label some OIDs as deprecated
We should avoid renumbering the OID namespace, as that could create
backwards compatilibity issues with existing NSDBs.  Fill in recently
introduced gaps in the FedFS OID namespace by adding back table rows
for the removed OIDs, and mark them deprecated.

For the new fedfsNceDN attribute, move it to a fresh OID, restore
the fedfsNcePrefix attribute in the OID namespace table, and mark
it deprecated.