Document FSN annotation and description attributes
The textual description of an FSL has a paragraph each for the
Annotation and Description attributes.  The textual description
of an FSN does not, even though an FSN can have these attributes.

It's also incorrect to say that "All FSLs have" annotations and
descriptions, when these are specified later in the draft as
_optional_ attributes.

Add appropriate descriptions for these attributes to the FSN
section.  Make the FSL description list these attributes as
optional.

I'm trying to distinguish carefully between the Fileset Name concept
and an "FSN record" which is how an FSN is represented in LDAP, as
the exact contents of the two are diverging.  Likewise for FSLs.