Partially address item NSDB-5
Review comment from

http://www.ietf.org/mail-archive/web/nfsv4/current/msg10851.html

| 4.2.1.14
|   XDR blobs? Really?

Section 4.2.1.14 describes the fedfsNfsPath attribute, part of the
NFS-specific subclass of FSL record types.  The value of this
attribute is converted by fileservers into part of an fs_locations4
or fs_locations_info4 response.  It is stored in LDAP as a binary
single-valued attribute.

Since the value of this attribute is already XDR-encoded, NFS
fileservers don't need to unmarshal and re-encode this content when
responding to a referral.  In addition, LDAP does not have a type
that can represent an ordered list of UTF-8 strings, so encoding of
some kind is required for this data item.

However, XDR blobs are binary.  Thus they are not human-readable
when stored in LDAP, and cannot be edited by typical LDAP clients.
We are interested in providing simple mechanisms for viewing and
updating FedFS content stored in LDAP.  A binary blob is not simple
to view or modify with typical LDAP tools.

LDAP experts and others are also not happy with the representation
of the fileserver network endpoint as a UTF-8 hostname string and a
separate port number.

To remedy this situation, replace the fedfsNfsPath, fedfsFslHost,
and fedfsFslPort attributes with a single attribute containing an
NFS URL that represents the fileset location encoded for this FSL
record.

Removing fedfsFslHost allows each subclass of fedFsFsl to specify
it's own way of representing the network location of a fileset.

There are several issues that must be resolved to make this a
successful change:

  o  NFS URLs are currently defined in RFC 2224 for NFS version 2
     and version 3.  This definition must be extended, within
     contemporary IETF guidelines for URIs (namely RFCs 3986 and
     4395), to accommodate NFS version 4.

  o  URLs are US-ASCII strings.  Since NFS pathnames and IDNA
     labels are UTF-8 strings, character encoding rules must be
     provided for both the hostname and pathname portion of an URL.

  o  URL pathnames use only forward slash as the pathname separator
     character.  Escape mechanisms must be provided to support
     alternate pathname separator characters.