Use the term "file-access client" consistently
My experience trying to explain FedFS to newcomers suggests we
should be more careful to disambiguate the use of the term "client."

The draft is already careful to use the term "NSDB client" or "LDAP
client" in all the appropriate places.  The term "file-access
client" is defined in the forward glossary, so let's use that term
consistently when referring to clients of fileservers.

Replace wholesale a couple of paragraphs due to inaccuracies or
other problems in the existing text.