Clarify discussion of mount points and junctions
I don't see the value in having separate definitions for "junction"
and "mount point."

According to the paragraph that defines mount points, these two
appear to be functionally the same thing.  This paragraph also
seems to imply that mount operations that construct the federated
namespace are performed by fileservers.

Finally, not all fileserver implementations use a "specially
marked directory" to implement a junction.  A symbolic link, for
example, can be used.