Remove "(o=fedfs by default)"
In my opinion, the default NCE DN has a bearing on interoperability,
especially since there is an NCE hunting mechanism.

What's more, the specified default NCE DN requires that a new
namingContext be set up on the NSDB, which is an additional
(perhaps minor) administrative burden for some LDAP servers.

We should leave the default NCE DN as an implementation choice.