Gen-ART review: Make glossary definitions consistent
Section 8, 3rd paragraph (definition of Client): "filesystem
access" -> "file-access" for consistency of usage with the rest of
the document.

Section 8, 5th paragraph (definition of Fileserver): rather than
discussing "a filesystem", should this be "one or more filesystems"?
Or is a fileserver limited to exporting one filesystem?

Section 8, 8th paragraph (definition of Filesystem Access Protocol):
following up on the 3rd paragraph, should this be "File-access
Protocol" for consistency?

This follows up on commit 8cd92fc1 'Use the term "file-access client"
consistently'.