Gen-ART review: LDAP-related corrections
Section 4.2.2: "LDAP Objects" -> "LDAP Object Classes" seems
appropriate.

Section 4.2.2.1, 2nd and 3rd sentences: replace "fedfsFsn" with
"fedfsNsdbContainerInfo"

Section 4.2.2.2, 5th paragraph: how is the prohibition on
referencing other attributes in the fedfsFsn object class supposed
to work if this document is the defining document for that object
class?

Section 5.1.3.1, 1st paragraph after LDIF definition: a port number
of 2049 is given.  Since this is already the default value, why not
use a different value?  Otherwise, there would be no practical need
to include that port number in the FSL creation request.