LDAP OID expert review: 'deprecated' -> 'historic'
The request provides the usage 'deprecated' for a number of
descriptors to be registered.  No such usage is described in BCP 64
nor is used in the registry.  It is unclear to me exactly what is
meant by this term as the I-D doesn't provide any guidance.
However, the registry does support marking registrations as being
'historic in nature'.  Assuming these descriptors are not used in
the specification and are otherwise truly historic in nature, I
would suggest replacing 'deprecated' with 'historic'.  Otherwise,
I suggest removing 'deprecated' from the registration requests.