IANA review: Clarify registry creation requests
So, should both new registries "FedFS Annotation Keys" and "FedFS Object
Identifiers" be existed in one single new top-level registry?  And if
so, what should the URL address be?  What is the title of the new
registry?  To which of the existing parameters listed at
http://www.iana.org/protocols FedFS registries do belonged to?
LDAP, NFSv4, or both?

URL: http://www.iana.org/assignments/???
Top-level Registry Title: NSDB Protocol for Federated Filesystems

Please revise the text in the new version to describe the above.