IESG review: Update glossary definitions of "client" and "fileserver"
Pete Resnick says:
> 1.1
>
>   Client:  Any client that accesses fileserver data using a supported
>      file-access protocol.
>
> I love circular definitions as much as the next guy, but how about at
> least indicating whether the client is the software, the machine, the
> user, or some other theoretical entity? The above is not useful.
>
>   Fileserver:  A server exporting one or more filesystems via a file-
>      access protocol.
>
> Hmmmm....That's not a terribly helpful definition.

The "glossary.xml" file is included in both the ADMIN and the NSDB
protocol documents, so the definitions are updated in both places.

Commit 8cd92fc1 'Use the term "file-access client" consistently',
Tue Sep 11 10:07:51 2012, changed the term "client" to "file-access
client" in the NSDB protocol document.  The same distinction should
be made when discussing ADMIN protocol operation status codes.