IESG review: Clarify i18n rules for FedFsPathComponent in ADMIN doc
Ralph Droms says:
>   FedFsPathComponent:  A case sensitive UTF-8 string containing a
>      filesystem path component.  It SHOULD be prepared using the
>      component4 rules defined in Chapter 12 "Internationalization" of
>      [3530bis].
>
> When would it not "be prepared using the component4 rules[...]" and
> how would that case be identified?

These component4 arrays are now specified as "utf8val_REQUIRED4" in
the XDR for the ADMIN protocol.

RFC 3530bis, section 12.2.2 says:

"utf8val_REQUIRED4/UVMUST:  indicating that strings of this type MUST
be in the form of the UTF-8 encoding of Unicode.  Strings will be
checked by the server for valid UTF-8 and the server SHOULD ensure
that when sent to the client, they are valid UTF-8."

Thus I'm changing the string preparation requirements for pathname
components in both the ADMIN and NSDB documents from "SHOULD be
prepared" to "MUST be prepared."