IESG review: Clarify meaning of "validity and consistency" (NSDB doc)
Sean Turner says:
> s5.1.1: When you say validity and consistency are talking about
> validating the schema?:
>
> The NSDB node that receives the request SHOULD check all of the
> attributes for validity and consistency

In addition to the document not being clear about what checking for
validity and consistency means, I don't see how not checking would be
an impedance to interoperability (ie, is an RFC 2119 qualifier like
SHOULD really appropriate here?).  In my opinion such validation
seems like an implementation quality issue, not an interoperability
issue.

In addition, the document ought to discuss how inconsistent or
invalid attribute values should be reported to LDAP clients, but
it does not.